Adds relayer.test.RelayerPatch
RelayerPatch is a patching utility so consumers of this library can easily
mock kafka communication when running tests.

This also refactors the tests of relayer to use RelayerPatch and updates
the documentation on its usage.
