+++ relayer/test/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from .relayer_patch import RelayerPatch
+from .mocked_producer import MockedProducer
