+++ tests/__init__.py
@@ -1,14 +1,19 @@
 import sure  # noqa
 from unittest import TestCase
 
-from .mocks import MockedProducer
+from relayer.test import RelayerPatch
 
 
 class BaseTestCase(TestCase):
-    def _setup_kafka_producer_mock(self, kafka_producer_mock):
-        mock_instance = MockedProducer()
-        kafka_producer_mock.return_value = mock_instance
-        self.producer = mock_instance
+
+    def setUp(self):
+        self.relayer_patch = RelayerPatch()
+        self.relayer_patch.start()
+        self.producer = self.relayer_patch.mocked_producer
+
+    def tearDown(self):
+        if hasattr(self, 'relayer_patch'):
+            self.relayer_patch.stop()
 
     def _get_topic_messages(self, topic):
         return self.producer.produced_messages[topic]
